Rename fake to demo.
